+++ gen/app/app.go
@@ -1,6 +1,8 @@
 package main
 
 import (
+	//"fmt"
+	//"github.com/viant/endly"
 	"fmt"
 	"github.com/viant/endly"
 	_ "github.com/viant/endly/gen/static"