+++ testing/dsunit/datastore/db1/schema.sql
@@ -1,0 +1,7 @@
+DROP TABLE IF EXISTS users;
+
+CREATE TABLE users (
+  id       INT AUTO_INCREMENT PRIMARY KEY,
+  email     VARCHAR(255) DEFAULT NULL
+);
+