+++ testing/runner/http/service_test.go
@@ -337,11 +337,11 @@ func Test_UdfProvider(t *testing.T) {
 	assert.NotNil(t, service)
 	context := manager.NewContext(toolbox.NewContext())
 	registerUDF, err := udf.NewRegisterRequestFromURL(path.Join(parentDir, "test/register_udf.json"))
-	if ! assert.Nil(t, err) {
+	if !assert.Nil(t, err) {
 		return
 	}
 	err = endly.Run(context, registerUDF, nil)
-	if ! assert.Nil(t, err) {
+	if !assert.Nil(t, err) {
 		return
 	}
 	request, err := runner.NewSendRequestFromURL(path.Join(parentDir, "test/udf_provider.json"))